Flytad koden från App.js
